+++ src/ui/dtl/actions/imgaug_augs/corruptlike/blur.md
