+++ src/ui/dtl/actions/imgaug_augs/corruptlike/compression.md
